Fix number of iterations (#1)
* Fix number of iterations 

This avoids different number of iterations due to changes in physics. So, only the computational power of the cluster is tested independent on geometry size

* Settings for benchmarking

* Disable surface function object

The surface function object uses parameters included from another file which does not work with OpenFOAM-7 running with multiple processes. It has been fixed in OpenFOAM-dev, see 
https://bugs.openfoam.org/view.php?id=3432